# music_app_backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  http.AddTrackToPlaylistRequest:
    properties:
      album:
        type: string
      artist:
        type: string
      artwork_url:
        type: string
      duration_ms:
        type: integer
      provider:
        type: string
      provider_track_id:
        type: string
      title:
        type: string
      track_number:
        type: integer
    required:
    - artist
    - provider
    - provider_track_id
    - title
    type: object
...
  http.CreatePlaylistRequest:
    properties:
      description:
        maxLength: 500
        type: string
      title:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - title
    type: object
...
  http.PlaylistResponse:
    properties:
      cover_url:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_public:
        type: boolean
      share_code:
        type: string
      title:
        type: string
      track_count:
        type: integer
      tracks:
        items:
          $ref: '#/definitions/http.PlaylistTrackResponse'
        type: array
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  http.PlaylistTrackResponse:
    properties:
      added_at:
        type: string
      album:
        type: string
      artist:
        type: string
      artwork_url:
        type: string
      duration_ms:
        type: integer
      id:
        type: string
      position:
        type: integer
      provider:
        type: string
      provider_track_id:
        type: string
      title:
        type: string
      track_number:
        type: integer
    type: object
...
  http.UpdatePlaylistRequest:
    properties:
      description:
        maxLength: 500
        type: string
      is_public:
        type: boolean
      title:
        maxLength: 100
        minLength: 3
        type: string
    type: object
...
  response.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  response.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.APIError'
      success:
        type: boolean
    type: object
  response.PaginatedData:
    properties:
      items: {}
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  response.SuccessMessage:
    properties:
      message:
        type: string
    type: object
host: localhost:8085
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the backend API for the Music App, providing endpoints for
    authentication, music discovery, playlists, and user library management.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Music App Backend API
  version: 1.2.0
paths:
...
  /playlists:
    get:
      description: Retrieves a paginated list of the authenticated user's playlists.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PaginatedData'
                  - properties:
                      playlists:
                        items:
                          $ref: '#/definitions/http.PlaylistResponse'
                        type: array
                    type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
      security:
      - Bearer: []
      summary: Get user playlists
      tags:
      - Playlists
    post:
      consumes:
      - application/json
      description: Creates a new playlist for the authenticated user.
      parameters:
      - description: Playlist details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.CreatePlaylistRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/http.PlaylistResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
      security:
      - Bearer: []
      summary: Create a new playlist
      tags:
      - Playlists
  /playlists/{playlistId}:
    delete:
      description: Permanently deletes a user's playlist.
      parameters:
      - description: Playlist ID
        in: path
        name: playlistId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.SuccessMessage'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
      security:
      - Bearer: []
      summary: Delete a playlist
      tags:
      - Playlists
    get:
      description: Retrieves details for a single playlist. Can be accessed without
        authentication if the playlist is public.
      parameters:
      - description: Playlist ID
        in: path
        name: playlistId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/http.PlaylistResponse'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
      security:
      - Bearer: []
      summary: Get a single playlist
      tags:
      - Playlists
    patch:
      consumes:
      - application/json
      description: Updates a playlist's title, description, or public status.
      parameters:
      - description: Playlist ID
        in: path
        name: playlistId
        required: true
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.UpdatePlaylistRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/http.PlaylistResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
      security:
      - Bearer: []
      summary: Update a playlist
      tags:
      - Playlists
  /playlists/{playlistId}/tracks:
    post:
      consumes:
      - application/json
      description: Adds a single track to the end of a specified playlist.
      parameters:
      - description: Playlist ID
        in: path
        name: playlistId
        required: true
        type: string
      - description: Track information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.AddTrackToPlaylistRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/http.PlaylistResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
      security:
      - Bearer: []
      summary: Add track to playlist
      tags:
      - Playlists
  /playlists/{playlistId}/tracks/{trackId}:
    delete:
      description: Removes a single track from a specified playlist.
      parameters:
      - description: Playlist ID
        in: path
        name: playlistId
        required: true
        type: string
      - description: Playlist Track ID
        in: path
        name: trackId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/http.PlaylistResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
      security:
      - Bearer: []
      summary: Remove track from playlist
      tags:
      - Playlists
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"